UCT/IB/MLX5: Fix invalid access error to strict-order key with offset
We access the atomic key from requester side assuming it was created
with offset, but actually that key was not created with atomic offset
(because KSM is not supported), and it's just a direct VA-based key.
This results in out-of-bounds memory access and RAE NAK.
The fix is for DevX memory domain to avoid using relaxed-order when
cannot create an indirect key with offset. Non-DevX memory domain sends
zero offset anyway.